word_book/requirements.txt
Django==1.11.2
djangorestframework==3.6.3
django-filter==1.0.4
markdown==2.6.8